refactor(router): create interfaces and move shared interfaces
